+++ paymentsheet/src/test/java/com/stripe/android/lpmfoundations/paymentmethod/definitions/IdealLpmBillingAddressTestCases.kt
@@ -4,8 +4,6 @@ import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadataFixt
 import com.stripe.android.model.Address
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
-import com.stripe.android.model.PaymentMethodExtraParams
-import com.stripe.android.model.PaymentMethodOptionsParams
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.uicore.elements.IdentifierSpec
 
@@ -22,11 +20,178 @@ private val idealSetupRawValues = idealRawValues + (
     IdentifierSpec.Email to "sanne.devries@example.com"
 )
 
-private val idealAutomaticWithoutTaxConfig = LpmBillingAddressTestConfiguration(
-    paymentMethodType = PaymentMethod.Type.Ideal,
-    billingDetailsCollectionMode = LpmBillingDetailsCollectionMode.AutomaticWithoutTax,
-    intentScenario = LpmBillingAddressTestConfiguration.IntentScenario.PaymentIntent,
-    termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
+private val idealNeverExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = null,
+        requiresMandate = false,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealAutomaticWithoutTaxExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = PaymentMethod.BillingDetails(
+            name = "Sanne de Vries",
+            address = Address(),
+        ),
+        requiresMandate = false,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+            "billing_details" to mapOf(
+                "name" to "Sanne de Vries",
+            ),
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealFullExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = PaymentMethod.BillingDetails(
+            name = "Sanne de Vries",
+            address = Address(
+                line1 = "Herengracht 1",
+                line2 = "Appartement 2",
+                city = "Amsterdam",
+                country = "NL",
+                postalCode = "1015BA",
+            ),
+        ),
+        requiresMandate = false,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+            "billing_details" to mapOf(
+                "name" to "Sanne de Vries",
+                "address" to mapOf(
+                    "line1" to "Herengracht 1",
+                    "line2" to "Appartement 2",
+                    "city" to "Amsterdam",
+                    "country" to "NL",
+                    "postal_code" to "1015BA",
+                ),
+            ),
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealNeverWithMandateExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = null,
+        requiresMandate = true,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealAutomaticWithoutTaxWithMandateExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = PaymentMethod.BillingDetails(
+            name = "Sanne de Vries",
+            email = "sanne.devries@example.com",
+            address = Address(),
+        ),
+        requiresMandate = true,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+            "billing_details" to mapOf(
+                "name" to "Sanne de Vries",
+                "email" to "sanne.devries@example.com",
+            ),
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealFullWithMandateExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = PaymentMethod.BillingDetails(
+            name = "Sanne de Vries",
+            email = "sanne.devries@example.com",
+            address = Address(
+                line1 = "Herengracht 1",
+                line2 = "Appartement 2",
+                city = "Amsterdam",
+                country = "NL",
+                postalCode = "1015BA",
+            ),
+        ),
+        requiresMandate = true,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+            "billing_details" to mapOf(
+                "name" to "Sanne de Vries",
+                "email" to "sanne.devries@example.com",
+                "address" to mapOf(
+                    "line1" to "Herengracht 1",
+                    "line2" to "Appartement 2",
+                    "city" to "Amsterdam",
+                    "country" to "NL",
+                    "postal_code" to "1015BA",
+                ),
+            ),
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
+)
+
+private val idealAutomaticWithoutTaxWithoutMandateExpectedParams = LpmBillingAddressFormParams(
+    createParams = PaymentMethodCreateParams.createWithOverride(
+        code = PaymentMethod.Type.Ideal.code,
+        billingDetails = PaymentMethod.BillingDetails(
+            name = "Sanne de Vries",
+            email = "sanne.devries@example.com",
+            address = Address(),
+        ),
+        requiresMandate = false,
+        overrideParamMap = mapOf(
+            "type" to PaymentMethod.Type.Ideal.code,
+            "billing_details" to mapOf(
+                "name" to "Sanne de Vries",
+                "email" to "sanne.devries@example.com",
+            ),
+        ),
+        productUsage = emptySet(),
+        allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
+        clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
+    ),
+    optionsParams = null,
+    extraParams = null,
 )
 
 internal val idealFormParamsTestCases = listOf(
@@ -39,31 +204,18 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = false,
-            params = emptyMap(),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = null,
-        ),
+        expectedParams = idealNeverExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL AutomaticWithoutTax PaymentIntent automatic terms",
-        config = idealAutomaticWithoutTaxConfig,
-        rawValues = idealRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = false,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
+        config = LpmBillingAddressTestConfiguration(
+            paymentMethodType = PaymentMethod.Type.Ideal,
+            billingDetailsCollectionMode = LpmBillingDetailsCollectionMode.AutomaticWithoutTax,
+            intentScenario = LpmBillingAddressTestConfiguration.IntentScenario.PaymentIntent,
+            termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
+        rawValues = idealRawValues,
+        expectedParams = idealAutomaticWithoutTaxExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL Full PaymentIntent automatic terms",
@@ -74,22 +226,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = false,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.address.line1" to "Herengracht 1",
-                "billing_details.address.line2" to "Appartement 2",
-                "billing_details.address.city" to "Amsterdam",
-                "billing_details.address.country" to "NL",
-                "billing_details.address.postal_code" to "1015BA",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealFullExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL Never PaymentIntentWithSetupFutureUsage automatic terms",
@@ -100,15 +237,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = emptyMap(),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = null,
-        ),
+        expectedParams = idealNeverWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL AutomaticWithoutTax PaymentIntentWithSetupFutureUsage automatic terms",
@@ -119,18 +248,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealAutomaticWithoutTaxWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL Full PaymentIntentWithSetupFutureUsage automatic terms",
@@ -141,23 +259,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-                "billing_details.address.line1" to "Herengracht 1",
-                "billing_details.address.line2" to "Appartement 2",
-                "billing_details.address.city" to "Amsterdam",
-                "billing_details.address.country" to "NL",
-                "billing_details.address.postal_code" to "1015BA",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealFullWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL Never SetupIntent automatic terms",
@@ -168,15 +270,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = emptyMap(),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = null,
-        ),
+        expectedParams = idealNeverWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL AutomaticWithoutTax SetupIntent automatic terms",
@@ -187,18 +281,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealAutomaticWithoutTaxWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL Full SetupIntent automatic terms",
@@ -209,23 +292,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.AUTOMATIC,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = true,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-                "billing_details.address.line1" to "Herengracht 1",
-                "billing_details.address.line2" to "Appartement 2",
-                "billing_details.address.city" to "Amsterdam",
-                "billing_details.address.country" to "NL",
-                "billing_details.address.postal_code" to "1015BA",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealFullWithMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL AutomaticWithoutTax PaymentIntentWithSetupFutureUsage never terms",
@@ -236,18 +303,7 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.NEVER,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = false,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealAutomaticWithoutTaxWithoutMandateExpectedParams,
     ),
     LpmBillingAddressFormValuesToParamsTestCase(
         name = "iDEAL AutomaticWithoutTax SetupIntent never terms",
@@ -258,94 +314,6 @@ internal val idealFormParamsTestCases = listOf(
             termsDisplay = PaymentSheet.TermsDisplay.NEVER,
         ),
         rawValues = idealSetupRawValues,
-        expectedParams = expectedFormParams(
-            type = PaymentMethod.Type.Ideal,
-            requiresMandate = false,
-            params = mapOf(
-                "billing_details.name" to "Sanne de Vries",
-                "billing_details.email" to "sanne.devries@example.com",
-            ),
-            optionsParams = null,
-            extraParams = null,
-            allowRedisplay = PaymentMethod.AllowRedisplay.UNSPECIFIED,
-            billingDetailsAddressWhenNoAddressParams = Address(),
-        ),
+        expectedParams = idealAutomaticWithoutTaxWithoutMandateExpectedParams,
     ),
 )
-
-internal fun expectedFormParams(
-    type: PaymentMethod.Type,
-    requiresMandate: Boolean,
-    params: Map<String, Any>,
-    optionsParams: PaymentMethodOptionsParams?,
-    extraParams: PaymentMethodExtraParams?,
-    allowRedisplay: PaymentMethod.AllowRedisplay,
-    billingDetailsAddressWhenNoAddressParams: Address?,
-): LpmBillingAddressFormParams {
-    val billingDetails = params.toBillingDetails(billingDetailsAddressWhenNoAddressParams)
-    val overrideParamMap = mapOf("type" to type.code) + params.unflatten()
-
-    return LpmBillingAddressFormParams(
-        createParams = PaymentMethodCreateParams.createWithOverride(
-            code = type.code,
-            billingDetails = billingDetails,
-            requiresMandate = requiresMandate,
-            overrideParamMap = overrideParamMap,
-            productUsage = emptySet(),
-            allowRedisplay = allowRedisplay,
-            clientAttributionMetadata = CLIENT_ATTRIBUTION_METADATA,
-        ),
-        optionsParams = optionsParams,
-        extraParams = extraParams,
-    )
-}
-
-private fun Map<String, Any>.toBillingDetails(
-    billingDetailsAddressWhenNoAddressParams: Address?,
-): PaymentMethod.BillingDetails? {
-    val billingDetailsPrefix = "billing_details."
-    if (keys.none { it.startsWith(billingDetailsPrefix) }) {
-        return null
-    }
-
-    val addressPrefix = "${billingDetailsPrefix}address."
-    val address = if (keys.any { it.startsWith(addressPrefix) }) {
-        Address(
-            city = get("${addressPrefix}city") as? String,
-            country = get("${addressPrefix}country") as? String,
-            line1 = get("${addressPrefix}line1") as? String,
-            line2 = get("${addressPrefix}line2") as? String,
-            postalCode = get("${addressPrefix}postal_code") as? String,
-            state = get("${addressPrefix}state") as? String,
-        )
-    } else {
-        billingDetailsAddressWhenNoAddressParams
-    }
-
-    return PaymentMethod.BillingDetails(
-        address = address,
-        email = get("${billingDetailsPrefix}email") as? String,
-        name = get("${billingDetailsPrefix}name") as? String,
-        phone = get("${billingDetailsPrefix}phone") as? String,
-    )
-}
-
-private fun Map<String, Any>.unflatten(): Map<String, Any> {
-    return entries.fold(linkedMapOf()) { result, (key, value) ->
-        val path = key.split(".")
-        var nestedMap: MutableMap<String, Any> = result
-        path.dropLast(1).forEach { segment ->
-            val child = nestedMap[segment]
-            if (child == null) {
-                val childMap = linkedMapOf<String, Any>()
-                nestedMap[segment] = childMap
-                nestedMap = childMap
-            } else {
-                @Suppress("UNCHECKED_CAST")
-                nestedMap = child as MutableMap<String, Any>
-            }
-        }
-        nestedMap[path.last()] = value
-        result
-    }
-}
